requirements.txt
discord.py>=2.3.2
python-dotenv
gspread
google-auth
aiohttp          # 헬스체크용 초간단 웹서버
